+++ smacc_sm_reference_library/sm_dance_bot/include/sm_dance_bot/states/st_event_count_down.h
@@ -1,3 +1,6 @@
+#include <smacc/smacc.h>
+namespace sm_dance_bot
+{
 struct StEventCountDown : smacc::SmaccState<StEventCountDown, MsDanceBotRunMode>
 {
     using SmaccState::SmaccState;
@@ -21,3 +24,4 @@ struct StEventCountDown : smacc::SmaccState<StEventCountDown, MsDanceBotRunMode>
         //static_createLogicUnit<LuEventCountdown, EvCountdownEnd<LuEventCountdown>, mpl::list<EvActionFeedback<smacc::SmaccMoveBaseActionClient>>>(100);
     }
 };
+}