+++ smacc_sm_reference_library/sm_dance_bot/include/sm_dance_bot/states/st_navigate_forward_2.h
@@ -1,3 +1,6 @@
+#include <smacc/smacc.h>
+namespace sm_dance_bot
+{
 struct StNavigateForward2 : smacc::SmaccState<StNavigateForward2, MsDanceBotRunMode>
 {
   typedef mpl::list<
@@ -24,3 +27,4 @@ struct StNavigateForward2 : smacc::SmaccState<StNavigateForward2, MsDanceBotRunM
   {
   }
 };
+}