+++ smacc_sm_reference_library/sm_dance_bot/include/sm_dance_bot/states/st_navigate_reverse_3.h
@@ -1,3 +1,6 @@
+#include <smacc/smacc.h>
+namespace sm_dance_bot
+{
 struct StNavigateReverse3 : smacc::SmaccState<StNavigateReverse3, MsDanceBotRunMode>
 {
    using SmaccState::SmaccState;
@@ -22,3 +25,4 @@ struct StNavigateReverse3 : smacc::SmaccState<StNavigateReverse3, MsDanceBotRunM
    {
    }
 };
+}