+++ smacc_sm_reference_library/sm_dance_bot/include/sm_dance_bot/states/st_rotate_degrees_2.h
@@ -1,3 +1,6 @@
+#include <smacc/smacc.h>
+namespace sm_dance_bot
+{
 struct StRotateDegrees2 : smacc::SmaccState<StRotateDegrees2, MsDanceBotRunMode>
 {
   using SmaccState::SmaccState;
@@ -23,3 +26,4 @@ struct StRotateDegrees2 : smacc::SmaccState<StRotateDegrees2, MsDanceBotRunMode>
     
   }
 };
+}