+++ smacc_sm_reference_library/sm_dance_bot/include/sm_dance_bot/states/st_rotate_degrees_4.h
@@ -1,3 +1,6 @@
+#include <smacc/smacc.h>
+namespace sm_dance_bot
+{
 struct StRotateDegrees4 : smacc::SmaccState<StRotateDegrees4, MsDanceBotRunMode>
 {
   using SmaccState::SmaccState;
@@ -22,3 +25,4 @@ struct StRotateDegrees4 : smacc::SmaccState<StRotateDegrees4, MsDanceBotRunMode>
   {
   }
 };
+}