+++ smacc_sm_reference_library/sm_dance_bot/include/sm_dance_bot/superstates/ss_radial_pattern_3.h
@@ -1,3 +1,5 @@
+#include <smacc/smacc.h>
+
 namespace sm_dance_bot
 {
 namespace SS3