+++ smacc_diagnostics/smacc_rviz_plugin/CHANGELOG.rst
@@ -5,7 +5,7 @@ Changelog for package smacc_rviz_plugin
 Forthcoming
 -----------
 * Progressing on transition log
-* progress on viewer, mode_states and history mode
+* progress on viewer, modestates and history mode
 * progress in the viewer and examples
 * Merge branch 'master' of https://github.com/reelrbtx/SMACC
 * Added SMACC thumbnail for RViz Plugin